[relay-runtime] Expand slightly and start splitting into files.
